requirements.txt
torch==1.6.0
matplotlib==3.1.3
pytorch_lightning==0.9.0rc16
torchvision==0.7.0
numpy==1.18.1
tqdm==4.42.1
Pillow==7.2.0
scikit_learn==0.23.2
